# modules/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bedrock Gateway API
  description: |
    Multi-tenant SaaS proxy for Amazon Bedrock with Cognito authentication,
    cascading budget management, rate limiting, and an admin dashboard.

    ## Authentication

    All API endpoints require authentication via one of:
    - **Cognito JWT Bearer Token**: `Authorization: Bearer <access_token>`
    - **Legacy Token Exchange**: `POST /auth/exchange` (deprecated)

    Human users authenticate via Cognito PKCE flow using the `bg-cognito-auth.sh` CLI.
    Agents/services use Cognito `client_credentials` flow for M2M authentication.

    ## Rate Limiting

    Rate limits are enforced at organization, department, team, and user levels.
    When rate limited, responses include:
    - `Retry-After` header
    - `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset` headers

    ## Budget Enforcement

    Budget limits cascade from organization to user level.
    When budget is exceeded (hard mode), requests are rejected with 429.
    Response headers include: `X-Budget-Remaining-USD`, `X-Budget-Period`
  version: 1.0.0
  contact:
    name: Platform Team
    email: platform@company.com
  license:
    name: Private
    url: https://company.com/license
servers:
  - url: https://{cloudfront_domain}/api
    description: Production Gateway via CloudFront
    variables:
      cloudfront_domain:
        default: gateway.company.com
        description: CloudFront distribution domain
...
security:
  - bearerAuth: []
...
paths:
...
  /usage/summary:
    get:
      tags: [usage]
      summary: Get usage summary
      operationId: getUsageSummary
      parameters:
        - name: org_id
          in: query
          schema:
            type: string
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
        - name: department_id
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Usage summary
...
  /usage/models:
    get:
      tags: [usage]
      summary: Get usage by model
      operationId: getUsageByModel
      parameters:
        - name: org_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Usage by model
  /usage/timeline:
    get:
      tags: [usage]
      summary: Get usage timeline
      operationId: getUsageTimeline
      parameters:
        - name: org_id
          in: query
          schema:
            type: string
        - name: interval
          in: query
          schema:
            type: string
            enum: [hourly, daily, weekly, monthly]
            default: daily
      responses:
        '200':
          description: Usage timeline
  /usage/users:
    get:
      tags: [usage]
      summary: Get usage by user
      operationId: getUsageByUser
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usage by user
  /usage/departments:
    get:
      tags: [usage]
      summary: Get usage by department
      operationId: getUsageByDepartment
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Usage by department
  /usage/logs:
    get:
      tags: [usage]
      summary: Get usage logs
      operationId: getUsageLogs
      parameters:
        - name: org_id
          in: query
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: model
          in: query
          schema:
            type: string
        - name: status_code
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Usage logs

  # ==========================================================================
  # Health Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Cognito JWT access token
...